Move Theano-dependent code in the core package
The following modules were moved into the garage.theano package:
- lasagne_layers
- lasagne_powered
- network

The corresponding imports were modified as well.

Also, the module lasagne_helpers was removed since no other module in
the project makes reference to it, with the functions in lasagne.layers
used instead.
