+++ garage/core/__init__.py
@@ -1,9 +1,2 @@
 from garage.core.serializable import Serializable
 from garage.core.parameterized import Parameterized  # noqa: I100
-from garage.core.lasagne_layers import batch_norm  # noqa: I100
-from garage.core.lasagne_layers import OpLayer
-from garage.core.lasagne_layers import ParamLayer
-from garage.core.lasagne_powered import LasagnePowered
-from garage.core.network import ConvNetwork
-from garage.core.network import GRUNetwork
-from garage.core.network import MLP
