+++ garage/policies/categorical_conv_policy.py
@@ -3,8 +3,6 @@ import lasagne.layers as L
 import lasagne.nonlinearities as NL
 import numpy as np
 
-from garage.core import ConvNetwork
-from garage.core import LasagnePowered
 from garage.core import Serializable
 from garage.distributions import Categorical
 from garage.envs.util import flatten, flatten_n, weighted_sample
@@ -13,6 +11,8 @@ from garage.misc import logger
 from garage.misc import tensor_utils
 from garage.misc.overrides import overrides
 from garage.policies import StochasticPolicy
+from garage.theano.core import ConvNetwork
+from garage.theano.core import LasagnePowered
 
 
 class CategoricalConvPolicy(StochasticPolicy, LasagnePowered):
