+++ garage/policies/gaussian_gru_policy.py
@@ -4,14 +4,14 @@ import lasagne.nonlinearities as NL
 import numpy as np
 import theano.tensor as TT
 
-from garage.core import GRUNetwork
-from garage.core import LasagnePowered
-from garage.core import ParamLayer
 from garage.core import Serializable
 from garage.distributions import RecurrentDiagonalGaussian
 from garage.misc import ext
 from garage.misc.overrides import overrides
 from garage.policies import StochasticPolicy
+from garage.theano.core import GRUNetwork
+from garage.theano.core import ParamLayer
+from garage.theano.core import LasagnePowered
 
 
 class GaussianGRUPolicy(StochasticPolicy, LasagnePowered):
