+++ garage/policies/gaussian_mlp_policy.py
@@ -4,9 +4,6 @@ import lasagne.nonlinearities as NL
 import numpy as np
 import theano.tensor as TT
 
-from garage.core import LasagnePowered
-from garage.core import MLP
-from garage.core import ParamLayer
 from garage.core import Serializable
 from garage.distributions import DiagonalGaussian
 from garage.misc import ext
@@ -14,6 +11,9 @@ from garage.misc import logger
 from garage.misc.overrides import overrides
 from garage.policies import StochasticPolicy
 from garage.spaces import Box
+from garage.theano.core import MLP
+from garage.theano.core import LasagnePowered
+from garage.theano.core import ParamLayer
 
 
 class GaussianMLPPolicy(StochasticPolicy, LasagnePowered):
