+++ garage/theano/core/__init__.py
@@ -1,0 +1,7 @@
+from garage.theano.core.lasagne_layers import batch_norm  # noqa: I100
+from garage.theano.core.lasagne_layers import OpLayer
+from garage.theano.core.lasagne_layers import ParamLayer
+from garage.theano.core.lasagne_powered import LasagnePowered
+from garage.theano.core.network import ConvNetwork
+from garage.theano.core.network import GRUNetwork
+from garage.theano.core.network import MLP
