+++ garage/theano/core/lasagne_layers.py
