+++ garage/theano/core/lasagne_powered.py
